refactor(engine): Rename db.models to db.schemas
