+++ tracecat/experimental/actions/core/cases.py
@@ -10,7 +10,7 @@ from pydantic import Field
 
 from tracecat.contexts import ctx_role
 from tracecat.db.engine import create_vdb_conn
-from tracecat.db.models import CaseContext
+from tracecat.db.schemas import CaseContext
 from tracecat.experimental.registry import registry
 from tracecat.types.api import Suppression, Tag
 from tracecat.types.cases import Case
